+++ R/extract.R
@@ -441,3 +441,74 @@ extract_as_tidygraph <- function(x, query = NULL, collapse_method = "first",
 
     tg
 }
+
+
+#' Extract mappings from ORDO
+#'
+#' Extract mappings from the Orphanet Rare Disease Ontology (ORDO). ORDO uses
+#' `oboInOwl:hasDbXref` for mapping with annotations to indicate
+#' exact/broad/narrow-ness. Utilizes [robot()].
+#'
+#' @param ordo_path The path to the ORDO file, as a string.
+#' @param as_skos Whether to convert ORDO's annotated `oboInOwl:hasDbXref`
+#' mappings to their
+#' [Simple Knowledge Organization System (SKOS)](https://www.w3.org/TR/2009/REC-skos-reference-20090818/)
+#' equivalents, as a boolean (default: `TRUE`).
+#'
+#' The ORDO-skos equivalent predicates are as follows:
+#'
+#' * `"BTNT"` - `skos:narrowMatch`
+#' * `"NTBT"` - `skos:broadMatch`
+#' * `"E"` - `skos:exactMatch`
+#' * `"ND"` - `doid:undefinedMatch` (supplements SKOS)
+#' * `"W"` - `doid:notMatch` (supplements SKOS)
+#'
+#' @param output The path where output will be written, as a string. If `NULL`
+#' (default), the data will be read into R and not saved to a file.
+#' @inheritDotParams tidy_sparql -query_res
+#' @returns
+#' If `output` is specified, the path to the output file with the data,
+#' otherwise, a `tibble` with the data.
+#'
+#' ORDO mappings data will be formatted according to the
+#' [SSSOM](https://github.com/mapping-commons/sssom) specification,
+#' with an additional `status` column indicating the status (active, deprecated,
+#' etc.) of each ORPHA term.
+#'
+#' If `as_skos = FALSE`, ORDO's text-based `oboInOwl:hasDbXref` annotations
+#' denoting the type of relationship the Xref represents (simple text code only)
+#' will be included in the `predicate_modifier` column.
+#'
+#' @export
+extract_ordo_mappings <- function(ordo_path, as_skos = TRUE, output = NULL,
+                                  ...) {
+    if (isTRUE(as_skos)) {
+        q_nm <- "mapping-ordo-skos.rq"
+    } else {
+        q_nm <- "mapping-ordo.rq"
+    }
+
+    q_file <- system.file("sparql", q_nm, package = "DO.utils", mustWork = TRUE)
+
+    if (is.null(output)) {
+        to_stdout <- TRUE
+        output <- tmp_out <- tempfile(fileext = ".tsv")
+        on.exit(unlink(tmp_out))
+    } else {
+        to_stdout <- FALSE
+    }
+
+    robot_query(input = ordo_path, query = q_file, output)
+
+    if (to_stdout) {
+        out <- readr::read_tsv(
+            output,
+            col_types = readr::cols(.default = readr::col_character())
+        )
+        out <- tidy_sparql(out, ...)
+    } else {
+        out <- output
+    }
+
+    out
+}